requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.5.7=h56e8100_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
cudatoolkit=11.8.0=h09e9e62_11
dataclasses-json=0.5.7=pypi_0
faiss=1.7.3=py310cuda112h2e38ea5_0_cuda
faiss-gpu=1.7.3=h949689a_0
filelock=3.12.0=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.5.0=pypi_0
greenlet=2.0.2=pypi_0
huggingface-hub=0.14.1=pypi_0
idna=3.4=pypi_0
intel-openmp=2023.1.0=h57928b3_46319
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
langchain=0.0.183=pypi_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libfaiss=1.7.3=cuda112h33bf9e0_0_cuda
libfaiss-avx2=1.7.3=cuda112h1234567_0_cuda
libffi=3.4.4=hd77b12b_0
liblapack=3.9.0=16_win64_mkl
llama-index=0.6.13=pypi_0
markupsafe=2.1.2=pypi_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
mkl=2022.1.0=h6a75c08_874
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
networkx=3.1=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.8.4=pypi_0
numpy=1.24.3=py310hd02465a_0
openai=0.27.7=pypi_0
openapi-schema-pydantic=1.2.4=pypi_0
openssl=1.1.1t=hcfcfb64_0
packaging=23.1=pypi_0
pandas=2.0.2=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=py310haa95532_0
pydantic=1.10.8=pypi_0
python=3.10.8=h966fe2a_1
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python_abi=3.10=2_cp310
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
regex=2023.5.5=pypi_0
requests=2.31.0=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=66.0.0=py310haa95532_0
six=1.16.0=pypi_0
sqlalchemy=2.0.15=pypi_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.12=pypi_0
tbb=2021.7.0=h91493d7_0
tenacity=8.2.2=pypi_0
threadpoolctl=3.1.0=pypi_0
tiktoken=0.4.0=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.13.3=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.65.0=pypi_0
transformers=4.29.2=pypi_0
typing-extensions=4.5.0=pypi_0
typing-inspect=0.8.0=pypi_0
tzdata=2023.3=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.16=pypi_0
vc=14.2=h21ff451_1
vc14_runtime=14.34.31931=h5081d32_16
vs2015_runtime=14.34.31931=hed1258a_16
wheel=0.38.4=py310haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
